requirements.txt
selenium==3.6.0
pytest-xdist===1.20.1